bugfix for base boundaries
When dist happens to be on exact base boundaries, the wrong decision was being made. For example when the distance between a and b is exactly 1, it will not correctly insert the second node. Leading to wrong size() computation and wrong neighbor retrieval.

Compare change to original algorithm which states <= instead of <.